apps/api/pyproject.toml
[project]
name = "fit-api"
version = "0.1.0"
description = "FitAI backend — FastAPI + PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.2",
    "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "httpx>=0.27.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
